+++ src/Janus.Hosting/IdentityEndpoints.cs
@@ -64,6 +64,7 @@ public static class IdentityEndpoints
         _ = library.MapPublication();
         _ = library.MapConfiguration();
         _ = library.MapRestrictions();
+        _ = library.MapDeliveryReports();
         _ = library.MapGrants();
         _ = library.MapRoles();
         _ = library.MapGroups();